Fix download token finder
Updated to get the token from the latest location. Allows for
downloading again. Shutterfly updated their system to use a diffrent
varible to store the ThisLife (shutterfly photos backend) server access
token. This update takes that into account and checks the new varible
for the access token to contiune downloading.